# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "XDFC 数字化管理 ERP 系统的 RESTful API 文档",
        "title": "XDFC 数字化管理 ERP API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@xdfc.com"
        },
        "license": {
            "name": "Proprietary",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "2.2.1"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/api/bom/{id}/tree-projection": {
            "get": {
                "description": "返回 BOM 的完整树投影数据，包括节点层级、折叠状态、元数据等",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "BOM"
                ],
                "summary": "获取 BOM 树投影",
                "parameters": [
                    {
                        "type": "string",
                        "description": "BOM ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "折叠空节点",
                        "name": "collapseEmpty",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "最大深度（0 = 无限制）",
                        "name": "maxDepth",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "已展开的节点 ID",
                        "name": "expandedNodeIds",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/services.TreeProjectionResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "BOM 不存在",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/maintenance-records": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "查询维保记录列表,支持分页、筛选、搜索",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "维保记录"
                ],
                "summary": "获取维保记录列表",
                "parameters": [
                    {
                        "enum": [
                            "MOLD",
                            "FURNACE"
                        ],
                        "type": "string",
                        "description": "资产类型",
                        "name": "assetType",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "资产ID",
                        "name": "assetId",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "OPEN",
                            "IN_PROGRESS",
                            "COMPLETED",
                            "CANCELLED"
                        ],
                        "type": "string",
                        "description": "状态",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "HIGH,CRITICAL",
                        "description": "优先级(逗号分隔)",
                        "name": "priority",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "PREVENTIVE",
                            "CORRECTIVE",
                            "INSPECTION"
                        ],
                        "type": "string",
                        "description": "维保类型",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "2026-01-01",
                        "description": "开始日期",
                        "name": "dateFrom",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "2026-12-31",
                        "description": "结束日期",
                        "name": "dateTo",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "搜索关键词(标题/序列号,最少2个字符)",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "maximum": 1000,
                        "type": "integer",
                        "default": 100,
                        "description": "每页数量",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "偏移量",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功返回记录列表",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "创建新的维保记录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "维保记录"
                ],
                "summary": "创建维保记录",
                "parameters": [
                    {
                        "description": "维保记录信息",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "创建成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/maintenance-records/stats": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "返回按状态分组的维保记录统计数据",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "维保记录"
                ],
                "summary": "获取维保记录统计",
                "responses": {
                    "200": {
                        "description": "统计数据",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/maintenance-records/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据ID获取维保记录详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "维保记录"
                ],
                "summary": "获取单条维保记录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "记录ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "维保记录详情",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "记录不存在",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "软删除维保记录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "维保记录"
                ],
                "summary": "删除维保记录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "记录ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "记录不存在",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "差分更新维保记录(SDRTS Delta 格式)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "维保记录"
                ],
                "summary": "更新维保记录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "记录ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新数据(Delta格式)",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "记录不存在",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "版本冲突",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "422": {
                        "description": "状态流转错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "services.TreeProjectionMetadata": {
            "type": "object",
            "properties": {
                "badge": {
                    "description": "徽章文本",
                    "type": "string"
                },
                "childCount": {
                    "description": "子节点数量",
                    "type": "integer"
                },
                "color": {
                    "description": "颜色",
                    "type": "string"
                },
                "icon": {
                    "description": "图标名称",
                    "type": "string"
                },
                "isEmpty": {
                    "description": "是否为空节点",
                    "type": "boolean"
                }
            }
        },
        "services.TreeProjectionNode": {
            "type": "object",
            "properties": {
                "branchRole": {
                    "description": "\"section\", \"collection\"",
                    "type": "string"
                },
                "childNodeIds": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "depth": {
                    "description": "节点深度（0 = root）",
                    "type": "integer"
                },
                "isCollapsed": {
                    "description": "是否已折叠",
                    "type": "boolean"
                },
                "isCollapsible": {
                    "description": "是否可折叠",
                    "type": "boolean"
                },
                "itemId": {
                    "description": "仅 leaf 节点",
                    "type": "string"
                },
                "label": {
                    "type": "string"
                },
                "metadata": {
                    "description": "元数据",
                    "allOf": [
                        {
                            "$ref": "#/definitions/services.TreeProjectionMetadata"
                        }
                    ]
                },
                "nodeId": {
                    "type": "string"
                },
                "nodeKind": {
                    "description": "\"root\", \"branch\", \"leaf\"",
                    "type": "string"
                },
                "parentNodeId": {
                    "type": "string"
                },
                "sectionCode": {
                    "type": "string"
                },
                "sectionName": {
                    "type": "string"
                }
            }
        },
        "services.TreeProjectionResponse": {
            "type": "object",
            "properties": {
                "nodes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/services.TreeProjectionNode"
                    }
                },
                "rootNodeId": {
                    "type": "string"
                },
                "totalCount": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT Token (格式: Bearer {token})",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}